requirements.txt
pillow
spotipy
requests

